LLMres/requirements.txt
azure-ai-formrecognizer==3.3.0
azure-core==1.29.5
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.11
langchain-openai==0.0.2
langchain-pinecone==0.0.1
openai>=1.6.1,<2.0.0
pinecone-client==3.0.2
python-dotenv==1.0.0
bs4==0.0.1
langchain-experimental>=0.0.49